note_project/requirements.txt
Flask
openai-whisper
transformers
torch
sumy